Full offline and install v1
Moved location types to json file in data folder as well as icon classes. They are also part of the service worker to cache the files for offline use of the app.
Removed styles and icons for weather app.
Moved glyphs into file in images.
Service worker is actually filled out to allow for cached/offline app.
All install code for installing the app has been completed.
perpetrated out rendering that is based on json objects we cache and load from server and static text that can be filled out right away.
Added separate divs for text and icons.