improved Dropbox integration
Separated url parsing into it's own function.
Add isMobile function to check if we are on mobile in js.
Now when the card is created the meta data is added to a dataset so it can be used in render.
Moved image display of dropbox images into render since offline we will just show a clickable link for now.
Cleaned up image display code and had it change preview links into downloadable for direct img tag using.
Hopefully fixed bugs with dropbox domains not being allowed (images not displaying, chooser not working...)