Got dropbox file saving working
Now you can save both the json and the text file to use on the map viewing site.
Updated the button to be white and a bit bigger.
Added a progress bar and text at top for time it happened.
Got a little nicer looking json and full text file formatting how the ink & parchment wants it.
I actually built the location formatting into the mincord-location class. May do the same for settings later.